refactor(codegen): extract std::math::* into stdlib/math.rs (Round 1.5)
4 lowering helpers (unary, nullary_float, is_nan, binary) lifted into
MathStdlib<'ctx> trait. The 17 std::math::* surface fns (sqrt, exp,
log, floor, ceil, pow, sin, cos, tan, asin, acos, atan, atan2, tanh,
nan, inf, is_nan) dispatch in codegen.rs picks the libm symbol and
delegates to one of these four trait helpers.

Surface bumps to pub(crate):
- Cx method: coerce_to_float (Int->Float widening helper, used widely)

codegen.rs: 44139 -> 43983 (-156 lines).
